Bug fix for IMU temperature

+++ src/imu/include/imu/LMS303.h
@@ -148,7 +148,7 @@ public:
 	int readFullSensorState();
 
 	int enableTempSensor();
-	int getTemperature();
+	float getTemperature();
 
 	int enableMagnetometer();
 	float getMagX() { return magX; }
